+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - Items should no longer be indicated as playing if the currently playing song is not contained
 within it
 - Fixed blurry playing indicator in album/artist/genre/playlist items
+- Fixed incorrect songs being displayed when adding albums to the end of the queue
 
 ## 3.1.0